seperate admin seeder in the db seed
